requirements.txt
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
confluent-kafka==2.3.0
google-api-core==2.15.0
google-auth==2.25.2
google-cloud-core==2.4.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.62.0
idna==3.4
kafka-python==2.0.2
numpy==1.26.1
pandas==2.1.2
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
six==1.16.0
tzdata==2023.3
urllib3==2.0.7
